CPack, Win32 compatibility work

+++ src/conf.c
@@ -44,7 +44,7 @@
 
 static int conf_readline(char **lineptr, size_t *n, FILE *stream)
 {
-	size_t so_far = 0;
+	int so_far = 0;
 	char temp;
 	char *temp_ptr;
